doesn't save, but looks nice

+++ E.P.C/Controllers/AIOsController.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using E.P.C.Data;
 using E.P.C.Models;
@@ -28,17 +27,10 @@ namespace E.P.C.Controllers
         // GET: AIOs/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
-            var aIO = await _context.AIOs
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (aIO == null)
-            {
-                return NotFound();
-            }
+            var aIO = await _context.AIOs.FirstOrDefaultAsync(m => m.Id == id);
+            if (aIO == null) return NotFound();
 
             return View(aIO);
         }
@@ -50,14 +42,15 @@ namespace E.P.C.Controllers
         }
 
         // POST: AIOs/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Brand,Model,RadiatorSize,PumpSpeed,FanSize,FanRPM,NoiseLevel,SupportedSockets,Id,Description,Price,ImageUrl")] AIO aIO)
+        public async Task<IActionResult> Create(
+            [Bind("Brand,Model,RadiatorSize,PumpSpeed,FanSize,FanRPM,NoiseLevel,Id,Description,Price,ImageUrl")] AIO aIO,
+            [FromForm] string[] SupportedSockets)
         {
             if (ModelState.IsValid)
             {
+                aIO.SupportedSockets = SupportedSockets != null ? string.Join(", ", SupportedSockets) : "";
                 _context.Add(aIO);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -68,48 +61,36 @@ namespace E.P.C.Controllers
         // GET: AIOs/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var aIO = await _context.AIOs.FindAsync(id);
-            if (aIO == null)
-            {
-                return NotFound();
-            }
+            if (aIO == null) return NotFound();
+
             return View(aIO);
         }
 
         // POST: AIOs/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Brand,Model,RadiatorSize,PumpSpeed,FanSize,FanRPM,NoiseLevel,SupportedSockets,Id,Description,Price,ImageUrl")] AIO aIO)
+        public async Task<IActionResult> Edit(
+            int id,
+            [Bind("Brand,Model,RadiatorSize,PumpSpeed,FanSize,FanRPM,NoiseLevel,Id,Description,Price,ImageUrl")] AIO aIO,
+            [FromForm] string[] SupportedSockets)
         {
-            if (id != aIO.Id)
-            {
-                return NotFound();
-            }
+            if (id != aIO.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
                 try
                 {
+                    aIO.SupportedSockets = SupportedSockets != null ? string.Join(", ", SupportedSockets) : "";
                     _context.Update(aIO);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!AIOExists(aIO.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!AIOExists(aIO.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -119,17 +100,10 @@ namespace E.P.C.Controllers
         // GET: AIOs/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
-            var aIO = await _context.AIOs
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (aIO == null)
-            {
-                return NotFound();
-            }
+            var aIO = await _context.AIOs.FirstOrDefaultAsync(m => m.Id == id);
+            if (aIO == null) return NotFound();
 
             return View(aIO);
         }
@@ -140,10 +114,7 @@ namespace E.P.C.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var aIO = await _context.AIOs.FindAsync(id);
-            if (aIO != null)
-            {
-                _context.AIOs.Remove(aIO);
-            }
+            if (aIO != null) _context.AIOs.Remove(aIO);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
